feat(enhancement): Allow NPCs to define cargo similarly to how Fleets can (#7113)
