+++ source/Fleet.h
@@ -16,6 +16,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #ifndef FLEET_H_
 #define FLEET_H_
 
+#include "FleetCargo.h"
 #include "Personality.h"
 #include "Sale.h"
 #include "Variant.h"
@@ -64,7 +65,8 @@ public:
 	void Enter(const System &system, std::list<std::shared_ptr<Ship>> &ships, const Planet *planet = nullptr) const;
 	// Place a fleet in the given system, already "in action." If the carried flag is set, only
 	// uncarried ships will be added to the list (as any carriables will be stored in bays).
-	void Place(const System &system, std::list<std::shared_ptr<Ship>> &ships, bool carried = true) const;
+	void Place(const System &system, std::list<std::shared_ptr<Ship>> &ships,
+			bool carried = true, bool addCargo = true) const;
 
 	// Do the randomization to make a ship enter or be in the given system.
 	// Return the system that was chosen for the ship to enter from.
@@ -78,7 +80,6 @@ private:
 	static std::pair<Point, double> ChooseCenter(const System &system);
 	std::vector<std::shared_ptr<Ship>> Instantiate(const std::vector<const Ship *> &ships) const;
 	bool PlaceFighter(std::shared_ptr<Ship> fighter, std::vector<std::shared_ptr<Ship>> &placed) const;
-	void SetCargo(Ship *ship) const;
 
 
 private:
@@ -87,10 +88,8 @@ private:
 	const Phrase *names = nullptr;
 	const Phrase *fighterNames = nullptr;
 	WeightedList<Variant> variants;
-	// The number of different items the ships in this fleet will carry in cargo.
-	int cargo = 3;
-	std::vector<std::string> commodities;
-	std::set<const Sale<Outfit> *> outfitters;
+	// The cargo ships in this fleet will carry.
+	FleetCargo cargo;
 
 	Personality personality;
 };
