+++ source/NPC.h
@@ -20,6 +20,7 @@ this program. If not, see <https://www.gnu.org/licenses/>.
 #include "EsUuid.h"
 #include "ExclusiveItem.h"
 #include "Fleet.h"
+#include "FleetCargo.h"
 #include "LocationFilter.h"
 #include "Personality.h"
 #include "Phrase.h"
@@ -104,6 +105,10 @@ private:
 	// The government of the ships in this NPC:
 	const Government *government = nullptr;
 	Personality personality;
+
+	// The cargo ships in this NPC will be able to carry.
+	FleetCargo cargo;
+	bool overrideFleetCargo = false;
 
 	EsUuid uuid;
